pyproject.toml
[project]
name = "fastapitut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asgiref>=3.11.1",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.128.6",
    "fastapi-mail>=1.6.2",
    "flower>=2.0.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "redis>=7.1.1",
    "schemathesis>=4.10.2",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
